+++ src/main/java/io/github/mzmine/modules/io/rawdataimport/fileformats/AgilentCsvReadTask.java
@@ -78,7 +78,7 @@ public class AgilentCsvReadTask extends AbstractTask {
           Range.closed(Double.parseDouble(range[0]), Double.parseDouble(range[1])));
       range = this.getMetaData(scanner, "time range").split(",");
       newMZmineFile.setRTRange(1,
-          Range.closed(Double.parseDouble(range[0]), Double.parseDouble(range[1])));
+          Range.closed((float) Double.parseDouble(range[0]), (float) Double.parseDouble(range[1])));
       totalScans = Integer.parseInt(this.getMetaData(scanner, "number of spectra"));
 
       // advance to the spectrum data...
@@ -93,7 +93,7 @@ public class AgilentCsvReadTask extends AbstractTask {
           return;
         } // if the task is canceled.
 
-        double retentionTime = scanner.nextDouble();
+        float retentionTime = (float) scanner.nextDouble();
         int msLevel = scanner.nextInt(); // not sure about this value
         scanner.next();
         scanner.next();
