+++ src/main/java/io/github/mzmine/modules/io/rawdataimport/fileformats/NetCDFReadTask.java
@@ -1,16 +1,16 @@
 /*
  * Copyright 2006-2020 The MZmine Development Team
- *
+ * 
  * This file is part of MZmine.
- *
+ * 
  * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
  * General Public License as published by the Free Software Foundation; either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
  * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public License along with MZmine; if not,
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
@@ -45,7 +45,7 @@ import ucar.nc2.NetcdfFile;
 import ucar.nc2.Variable;
 
 /**
- *
+ * 
  */
 public class NetCDFReadTask extends AbstractTask {
 
@@ -379,7 +379,7 @@ public class NetCDFReadTask extends AbstractTask {
     scanLength[0] = startAndLength[1];
 
     // Get retention time of the scan
-    Double retentionTime = scansRetentionTimes.get(scanNum);
+    Float retentionTime = scansRetentionTimes.get(scanNum).floatValue();
     if (retentionTime == null) {
       logger.severe("Could not find retention time for scan " + scanNum);
       throw (new IOException("Could not find retention time for scan " + scanNum));
@@ -389,7 +389,7 @@ public class NetCDFReadTask extends AbstractTask {
     if (scanLength[0] == 0) {
       scanNum++;
 
-      return new SimpleScan(null, scanNum, 1, retentionTime.doubleValue(),
+      return new SimpleScan(null, scanNum, 1, retentionTime,
           0, 0, null, new DataPoint[0], MassSpectrumType.CENTROIDED,
           PolarityType.UNKNOWN, "", null);
     }
@@ -434,7 +434,7 @@ public class NetCDFReadTask extends AbstractTask {
     // Auto-detect whether this scan is centroided
     MassSpectrumType spectrumType = ScanUtils.detectSpectrumType(dataPoints);
 
-    SimpleScan buildingScan = new SimpleScan(null, scanNum, 1, retentionTime.doubleValue(),
+    SimpleScan buildingScan = new SimpleScan(null, scanNum, 1, retentionTime,
         0, 0, null, dataPoints, spectrumType, polarity,
         scanDefinition, null);
