+++ src/main/java/io/github/mzmine/project/impl/StorableScan.java
@@ -50,7 +50,7 @@ public class StorableScan implements Scan {
   private int scanNumber, msLevel, fragmentScans[];
   private double precursorMZ;
   private int precursorCharge;
-  private double retentionTime;
+  private float retentionTime;
   private Range<Double> mzRange;
   private DataPoint basePeak;
   private Double totalIonCurrent;
@@ -96,7 +96,7 @@ public class StorableScan implements Scan {
   }
 
   public StorableScan(RawDataFileImpl rawDataFile, int storageID, int numberOfDataPoints,
-      int scanNumber, int msLevel, double retentionTime, double precursorMZ,
+      int scanNumber, int msLevel, float retentionTime, double precursorMZ,
       int precursorCharge, int fragmentScans[], MassSpectrumType spectrumType,
       PolarityType polarity, String scanDefinition, Range<Double> scanMZRange) {
 
@@ -106,7 +106,7 @@ public class StorableScan implements Scan {
   }
 
   public StorableScan(RawDataFileImpl rawDataFile, int storageID, int numberOfDataPoints,
-      int scanNumber, int msLevel, double retentionTime, double precursorMZ,
+      int scanNumber, int msLevel, float retentionTime, double precursorMZ,
       int precursorCharge, int fragmentScans[], MassSpectrumType spectrumType,
       PolarityType polarity, String scanDefinition, Range<Double> scanMZRange, double mobility,
       MobilityType mobilityType) {
@@ -255,7 +255,7 @@ public class StorableScan implements Scan {
    * @see io.github.mzmine.datamodel.Scan#getScanAcquisitionTime()
    */
   @Override
-  public double getRetentionTime() {
+  public float getRetentionTime() {
     return retentionTime;
   }
